deployer now passes on any defined env vars
